requirements.txt
amqp==5.1.1
appdirs==1.4.4
arabic-reshaper==3.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
asn1crypto==1.5.1
async-timeout==4.0.2
audioread==2.1.9
autopep8==1.5.6
bcrypt==3.2.0
billiard==3.6.4.0
boto==2.49.0
boto3==1.21.3
botocore==1.24.3
captcha==0.3
celery==5.2.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cryptography==39.0.2
cssselect2==0.7.0
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.14
django-cleanup==6.0.0
django-cors-headers==3.11.0
django-import-export==2.7.1
django-storages==1.12.3
django-utils-six==2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
et-xmlfile==1.1.0
fonttools==4.39.0
fpdf2==2.6.1
gunicorn==20.0.4
html5lib==1.1
idna==3.3
jmespath==0.10.0
joblib==1.1.0
kombu==5.2.4
librosa==0.9.1
llvmlite==0.38.0
lxml==4.9.2
MarkupPy==1.14
numba==0.55.1
numpy==1.21.5
odfpy==1.4.1
openpyxl==3.0.9
oscrypto==1.3.0
packaging==21.3
Pillow==9.0.1
pooch==1.6.0
prompt-toolkit==3.0.38
psycopg2==2.9.3
pycodestyle==2.7.0
pycparser==2.21
pyHanko==0.17.0
pyhanko-certvalidator==0.20.1
PyJWT==2.3.0
pyparsing==3.0.7
pypdf==3.5.1
pypng==0.0.21
PyQRCode==1.2.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
redis==4.5.1
reportlab==3.6.12
requests==2.27.1
resampy==0.2.2
s3transfer==0.5.1
scikit-learn==1.0.2
scipy==1.8.0
shortuuid==1.0.11
six==1.16.0
SoundFile==0.10.3.post1
sqlparse==0.4.2
svglib==1.5.1
tablib==3.2.0
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
typing_extensions==4.5.0
tzdata==2021.5
tzlocal==4.2
uritools==4.0.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
wincertstore==0.2
xhtml2pdf==0.2.9
xlrd==2.0.1
xlwt==1.3.0
